[FIX] #68 - Retrofit API 호출 에러 처리
- Retrofit API를 호출하는 부분을 try-catch로 감싸서 에러 핸들링
