+++ app/src/main/java/com/example/android_repo_05/data/repositories/ProfileImageRepository.kt
@@ -1,6 +1,11 @@
 package com.example.android_repo_05.data.repositories
 
+import com.example.android_repo_05.data.models.ResponseState
+import com.example.android_repo_05.data.models.UserProfileResponse
 import com.example.android_repo_05.retrofit.GithubApiInstance
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import retrofit2.Response
 
 class ProfileImageRepository {
     companion object {
@@ -13,7 +18,20 @@ class ProfileImageRepository {
         }
     }
 
-    suspend fun getProfileImageFromRemote() =
-        GithubApiInstance.retrofit.getProfileUrl()
+    suspend fun getProfileImageFromRemote() = withContext(Dispatchers.IO) {
+        return@withContext try {
+            handleLoginResponse(GithubApiInstance.retrofit.getProfileUrl())
+        } catch (e: Exception) {
+            ResponseState.Error(e.message ?: "error")
+        }
+    }
 
+    private fun handleLoginResponse(response: Response<UserProfileResponse>): ResponseState<UserProfileResponse> {
+        if (response.isSuccessful) {
+            response.body()?.let { result ->
+                return ResponseState.Success(result)
+            }
+        }
+        return ResponseState.Error(response.message())
+    }
 }
